Refactor: Split into separate files
