+++ src/main.ts
@@ -1,253 +1,9 @@
 import { setup, createActor, fromPromise, assign } from "xstate";
-import * as readline from 'readline';
-// Imports for server and file system (required for audio importing)
-import * as http from 'http';
-import * as fs from 'fs';
-import * as path from 'path';
-
-import { networkInterfaces } from 'os';
-
-
-// Automatically find your computer's IP
-function findIPContaining(value: string) {
-  for (const [name, ifaceArray] of Object.entries(networkInterfaces())) {
-    if (!ifaceArray) continue;
-    for (const iface of ifaceArray) {
-      if (iface.address.indexOf(value) === 0) {
-        return iface.address;
-      }
-    }
-  }
-}
-
-const OVERRIDE_IP = ""; // If automatic detection doesn't work, set here. Run ifconfig | grep 192 in terminal to find it. Or ifconfig | grep 10
-
-const PC_IP = OVERRIDE_IP || findIPContaining("192") || findIPContaining("10."); // CHANGE THIS TO YOUR COMPUTER'S LOCAL IP ADDRESS (the one that Furhat can access through the network). Run ifconfig | grep 192 on terminal to find it. It usually starts with 192.168.1.xxx or 10.0.0.xxx
-const FURHATURI = "192.168.1.11:54321";
-const OLLAMA_API_URL = "http://localhost:11434/api/chat";
-const firstMessageWaitTimeMs = 0; // 0 second for the first message
-
-// >>> HANDLING RECORDED AUDIO IMPORTS /BEGIN <<<
-const AUDIO_PORT = 8000;
-const AUDIO_DIR = path.join(__dirname, 'audio');
-
-http.createServer((req, res) => {
-  const filePath = path.join(AUDIO_DIR, req.url || '');
-  
-  if (fs.existsSync(filePath) && filePath.endsWith('.wav')) {
-    res.writeHead(200, { 'Content-Type': 'audio/wav' });
-    fs.createReadStream(filePath).pipe(res);
-  } else {
-    res.writeHead(404);
-    res.end('File not found');
-  }
-}).listen(AUDIO_PORT, () => {
-  console.log(`Audio server running at http://${PC_IP}:${AUDIO_PORT}`);
-});
-
-const audioFiles: Record<string, string> = {
-  '1': `http://${PC_IP}:${AUDIO_PORT}/hmm_doctor.wav`,
-  '2': `http://${PC_IP}:${AUDIO_PORT}/hmm_pregnant.wav`,
-  '3': `http://${PC_IP}:${AUDIO_PORT}/hmm_child.wav`,
-  '4': `http://${PC_IP}:${AUDIO_PORT}/hmm_pilot.wav`,
-
-  'q': `http://${PC_IP}:${AUDIO_PORT}/pause_doctor.wav`,
-  'w': `http://${PC_IP}:${AUDIO_PORT}/pause_pregnant.wav`,
-  'e': `http://${PC_IP}:${AUDIO_PORT}/pause_child.wav`,
-  'r': `http://${PC_IP}:${AUDIO_PORT}/pause_pilot.wav`,
-
-  'a': `http://${PC_IP}:${AUDIO_PORT}/hahaha_doctor.wav`,
-  's': `http://${PC_IP}:${AUDIO_PORT}/hahaha_pregnant.wav`,
-  'd': `http://${PC_IP}:${AUDIO_PORT}/hahaha_child.wav`,
-  'f': `http://${PC_IP}:${AUDIO_PORT}/hahaha_pilot.wav`,
-};
-// >>> HANDLING RECORDED AUDIO IMPORTS /END <<<
-
-// Types
-type Message = { // LLM dialogue structure. The system will constantly change between these roles at each turn.
-  role: "assistant" | "user" | "system"; // system is a sole actor. Assistant is the LLM. User is us.
-  content: string;
-};
-
-interface DMContext { // Our regular DMContext types.
-  lastResult: string;
-  messages: Message[];
-  isFirstMessage: boolean; // If the message is the first message.
-  pendingManipulation: string | null; // Stores the manipulation phrase to add to next assistant turn
-  keyPressed: string | null; // Stores which key was pressed
-  userSpeechBuffer: string[]; // NEW: Accumulates user utterances before processing
-}
-
-// Setup readline interface for keyboard input in Node.js
-readline.emitKeypressEvents(process.stdin);
-if (process.stdin.isTTY) {
-  process.stdin.setRawMode(true);
-}
-
-// Furhat API functions
-async function fhVoice(name: string) { // fh functions are fetched from Furhat's URI. They are ready-made functions.
-  const myHeaders = new Headers();
-  myHeaders.append("accept", "application/json");
-  const encName = encodeURIComponent(name);
-  return fetch(`http://${FURHATURI}/furhat/voice?name=${encName}`, {
-    method: "POST",
-    headers: myHeaders,
-    body: "",
-  });
-}
-
-async function fhSay(text: string, isFirstMessage: boolean = false) { 
-  const myHeaders = new Headers();
-  myHeaders.append("accept", "application/json");
-  const encText = encodeURIComponent(text);
-  await fetch(`http://${FURHATURI}/furhat/say?text=${encText}&blocking=true`, {
-    method: "POST",
-    headers: myHeaders,
-    body: "",
-  });
-  
-  // 6 second delay for first message (long introduction), 1 second for others
-  const delay = isFirstMessage ? firstMessageWaitTimeMs : 200;
-  await new Promise(resolve => setTimeout(resolve, delay));
-}
-
-const timer = fromPromise(
-  ({ input }: { input: { ms: number } }) =>
-    new Promise((resolve) => setTimeout(resolve, input.ms))
-);
-
-async function fhSayAudio(audioUrl: string, isFirstMessage: boolean = false) {
-  const myHeaders = new Headers();
-  myHeaders.append("accept", "application/json");
-  const encUrl = encodeURIComponent(audioUrl);
-  
-  // Remove the 'text=' and use 'url=' instead
-  await fetch(`http://${FURHATURI}/furhat/say?url=${encUrl}&blocking=true&lipsync=true`, {
-    method: "POST",
-    headers: myHeaders,
-    body: "",
-  });
-  
-  const delay = isFirstMessage ? firstMessageWaitTimeMs : 200;
-  await new Promise(resolve => setTimeout(resolve, delay));
-}
-
-async function fhAttendUser() { // This is about GAZE.
-  const myHeaders = new Headers();
-  myHeaders.append("accept", "application/json");
-  return fetch(`http://${FURHATURI}/furhat/attend?user=CLOSEST`, { // Look at documentation (https://docs.furhat.io/remote-api/) in the "Attend" section
-    /*
-    # Attend the user closest to the robot
-    furhat.attend(user="CLOSEST") 
-
-    There are other attend options in the doc.
-    */
-    method: "POST",
-    headers: myHeaders,
-    body: "",
-  });
-}
-
-async function fhListen(): Promise<string> { // Furhat's own ASR.
-  const myHeaders = new Headers();
-  myHeaders.append("accept", "application/json");
-
-  // Before listenıng, we send a request to stop listening in case Furhat is still processing previous audio. This is a workaround to prevent the problem of Furhat not responding after the first turn due to some issue with the listen endpoint. After sending the stop command, we immediately send the listen command again to start listening for new input.
-  return fetch(`http://${FURHATURI}/furhat/listen/stop`, {
-    method: "POST",
-    headers: myHeaders,
-  }).then(() => {
-    console.log("(Re)starting to listen...");
-    return fetch(`http://${FURHATURI}/furhat/listen`, {
-      method: "GET",
-      headers: myHeaders,
-    });
-  })
-    .then((response) => response.body)
-    .then((body) => body!.getReader().read())
-    .then((reader) => reader.value)
-    .then((value) => JSON.parse(new TextDecoder().decode(value!)).message);
-}
-
-// Ollama API function
-async function fetchChatCompletion(messages: Message[]): Promise<string> {
-  console.log("Calling Ollama with messages:", messages);
-  
-  try {
-    const response = await fetch(OLLAMA_API_URL, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        model: "llava:13b",
-        messages: messages,
-        stream: false,
-      }),
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      console.error("Ollama API error:", response.status, errorText);
-      throw new Error(`Ollama API error: ${response.status}`);
-    }
-    
-    const data = await response.json();
-    console.log("Ollama response:", data);
-    
-    const assistantMessage = data.message.content;
-    return assistantMessage;
-  } catch (error) {
-    console.error("Error calling Ollama:", error);
-    return "Error while connecting to the language model. Probably ssh tunnel is not active.";
-  }
-}
-
-// Keyboard input listener for Node.js - waits for a single keypress and returns the key
-async function waitForKeypress(): Promise<string> {
-  return new Promise((resolve) => {
-    const handler = (str: string, key: any) => {
-      // Handle Ctrl+C to exit gracefully
-      if (key.ctrl && key.name === 'c') {
-        console.log('\nExiting...');
-        process.exit(0);
-      }
-      
-      resolve(key.name.toLowerCase());
-    };
-    process.stdin.removeAllListeners('keypress');
-    process.stdin.once('keypress', handler);
-  });
-}
-
-const laughterKeys = ['a', 's', 'd', 'f'];
-const pauseKeys = ['q', 'w', 'e', 'r'];
-const filledPauseKeys = ['1', '2', '3', '4'];
-const hypothesisKeys = [...laughterKeys, ...pauseKeys, ...filledPauseKeys];
-
-const switchTopicKeys = ['z', 'x', 'c', 'v'];
-const forceConcludeKey = 'b';
-const nextTopicKey = 'n';
-const guidanceKeys = [...switchTopicKeys, forceConcludeKey, nextTopicKey];
-
-const allManipulationKeys = [...hypothesisKeys, ...guidanceKeys];
-
-// NEW: Combined actor that races between listening and waiting for keypress
-const listenOrKeypress = fromPromise(async () => {
-  return Promise.race([
-    fhListen().then(result => ({ type: 'speech' as const, data: result })),
-    waitForKeypress().then(result => ({ type: 'keypress' as const, data: result }))
-  ]);
-});
-
-// NEW: Helper function to filter out NOMATCH and replace with "..."
-function sanitizeUtterance(utterance: string): string {
-  // Check if the utterance contains "NOMATCH" (case-insensitive)
-  if (utterance.toLowerCase().includes('nomatch')) {
-    return '...';
-  }
-  return utterance;
-}
+import { Message, DMContext } from "./types";
+import { timer, fhVoice, fhAttendUser, fhSay, fhSayAudio, fhListen, sanitizeUtterance } from "./furhat";
+import { fetchChatCompletion } from "./ollama";
+import { waitForKeypress, listenOrKeypress } from "./keypress";
+import { allManipulationKeys, hypothesisKeys, manipulations, audioFiles } from "./audio_manipulations";
 
 // State machine
 const dmMachine = setup({
@@ -549,42 +305,14 @@ const dmMachine = setup({
     // Add manipulation phrase based on key pressed
     AddManipulation: {
       entry: assign(({ context }) => {
-        // Map keys to manipulation phrases
-        const manipulations: Record<string, string> = {
-          // Hmm interventions (1-4) -- audio cued
-          '1': `http://${PC_IP}:${AUDIO_PORT}/hmm_doctor.wav`,
-          '2': `http://${PC_IP}:${AUDIO_PORT}/hmm_pregnant.wav`,
-          '3': `http://${PC_IP}:${AUDIO_PORT}/hmm_child.wav`,
-          '4': `http://${PC_IP}:${AUDIO_PORT}/hmm_pilot.wav`,
-          // Pause versions (q, w, e, r) -- audio cued
-          'q': `http://${PC_IP}:${AUDIO_PORT}/pause_doctor.wav`,
-          'w': `http://${PC_IP}:${AUDIO_PORT}/pause_pregnant.wav`,
-          'e': `http://${PC_IP}:${AUDIO_PORT}/pause_child.wav`,
-          'r': `http://${PC_IP}:${AUDIO_PORT}/pause_pilot.wav`,
-          // Hahaha interventions (a, s, d, f) -- audio cued
-          'a': `http://${PC_IP}:${AUDIO_PORT}/hahaha_doctor.wav`,
-          's': `http://${PC_IP}:${AUDIO_PORT}/hahaha_pregnant.wav`,
-          'd': `http://${PC_IP}:${AUDIO_PORT}/hahaha_child.wav`,
-          'f': `http://${PC_IP}:${AUDIO_PORT}/hahaha_pilot.wav`,
-          // Switch topic interventions (z, x, c, v) -- direct text manipulation
-          'z': 'Cool, shall we talk about the doctor now?',
-          'x': 'Great, shall we talk about the pregnant lady now?',
-          'c': 'Perfect, shall we talk about the child now?',
-          'v': 'Nice, shall we talk about the pilot now?',
-          // Switch to the next topic -- direct text manipulation
-          'n': 'Good, shall we talk about the next passenger?',
-          // Force conclusion -- direct text manipulation
-          'b': 'So, based on your discussions, who do you think should jump?',
-        };
-
         // The following if statement checks if the key pressed is one of the hypothesis manipulation keys (1-4, q-r, a-f). If so, it queues an audio manipulation. 
         // Otherwise, it adds a text manipulation phrase for guidance keys (z-v, n, b).
-      if (hypothesisKeys.includes(context.keyPressed || '')) { 
+        if (hypothesisKeys.includes(context.keyPressed || '')) { 
           const audioUrl = audioFiles[context.keyPressed || ''];
           const textForHistory = `[Audio manipulation: ${context.keyPressed}]`;
           
           console.log(`Queuing audio: ${audioUrl}`);
-          
+
           return {
             messages: [
               ...context.messages,
